The one with the UI.
Added a UI class that builds a jframe, which the user will interface
with. Also added a facade that acts as a link between the UI and the
subsystem underneath. GlyphBuilder was updated to include a clear
method. So far, the system can load a file (called "text.txt") and build
a tree from that, or can build a new tree. Every time the user makes a
change in the text pane, the update button must be pressed to reflect
the changes in the tree structure.